refactor: fix a myriad of things
